Code reorganization and little and implementation of --run option
